requirements.txt
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
dlib==19.24.4
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
parso==0.8.4
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
wcwidth==0.2.13
wheel==0.43.0
